AI_python/requirements.txt
h5py==3.1.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.18
numpy==1.19.5
pandas==1.3.2
Pillow==8.3.1
plaidml-keras==0.7.0
PyMySQL==1.0.2
PyYAML==5.4.1
scikit-image==0.18.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
six==1.15.0
imageio==2.9.0
plaidml==0.7.0
tifffile==2021.10.10
